Refactored the build and include paths.
This tool did not use the include paths associated with a standard
htslib install. It also did not do this for samtools, but this is
because samtools did not install as a library. This uses a samtools fork
(for now) that has library install support.

The dockerfile now builds root from source, and installs the
dependencies in the standard systems paths (or a path prefix specifying
another location). This makes the builds much more simpler and workable
with other projects. Also, the docker image uses a smaller base image as
well as a mutl-stage build to shrink the size and minimize the surface
of the image. This image could be further shrunk by only copying exactly
the compiled library files needed to link to cnvnator.